[BEAM-5985] Extracted common part for Nexmark and LoadTests in Jenkins configuration. Some fixes.
