+++ .test-infra/jenkins/job_LoadTests_Java.groovy
@@ -16,9 +16,9 @@
  * limitations under the License.
  */
 
-import LoadTestsBuilder as loadTestsBuilder
 import CommonJobProperties as commonJobProperties
-
+import LoadTestsBuilder as loadTestsBuilder
+import PhraseTriggeringPostCommitBuilder
 
 def testsConfigurations = [
         [
@@ -26,8 +26,8 @@ def testsConfigurations = [
                 jobDescription    : 'Runs GroupByKey load tests on direct runner small records 10b',
                 itClass           : 'org.apache.beam.sdk.loadtests.GroupByKeyLoadTest',
                 prCommitStatusName: 'Java GroupByKey Small Java Load Test Direct',
-                prTriggerPhase    : 'Run GroupByKey Small Java Load Test Direct',
-                runner            : loadTestsBuilder.Runner.DIRECT,
+                prTriggerPhrase   : 'Run GroupByKey Small Java Load Test Direct',
+                runner            : CommonTestProperties.Runner.DIRECT,
                 jobProperties     : [
                         publishToBigQuery: true,
                         bigQueryDataset  : 'load_test_PRs',
@@ -44,8 +44,8 @@ def testsConfigurations = [
                 jobDescription    : 'Runs GroupByKey load tests on Dataflow runner small records 10b',
                 itClass           : 'org.apache.beam.sdk.loadtests.GroupByKeyLoadTest',
                 prCommitStatusName: 'Java GroupByKey Small Load Test Dataflow',
-                prTriggerPhase    : 'Run GroupByKey Small Java Load Test Dataflow',
-                runner            : loadTestsBuilder.Runner.DATAFLOW,
+                prTriggerPhrase   : 'Run GroupByKey Small Java Load Test Dataflow',
+                runner            : CommonTestProperties.Runner.DATAFLOW,
                 jobProperties     : [
                         publishToBigQuery   : true,
                         bigQueryDataset     : 'load_test_PRs',
@@ -61,25 +61,14 @@ def testsConfigurations = [
 ]
 
 for (testConfiguration in testsConfigurations) {
-    create_load_test_job(testConfiguration)
-}
-
-private void create_load_test_job(testConfiguration) {
-
-    // This job runs load test with Metrics API
-    job(testConfiguration.jobName) {
+    PhraseTriggeringPostCommitBuilder.postCommitJob(
+            testConfiguration.jobName,
+            testConfiguration.prTriggerPhrase,
+            testConfiguration.prCommitStatusName,
+            this
+    ) {
         description(testConfiguration.jobDescription)
-
-        // Set default Beam job properties.
         commonJobProperties.setTopLevelMainJobProperties(delegate)
-
-        // Allows triggering this build against pull requests.
-        commonJobProperties.enablePhraseTriggeringFromPullRequest(
-                delegate,
-                testConfiguration.prCommitStatusName,
-                testConfiguration.prTriggerPhase)
-
-
         loadTestsBuilder.buildTest(delegate, testConfiguration.jobDescription, testConfiguration.runner, testConfiguration.jobProperties, testConfiguration.itClass)
     }
 }
